+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v1.0.0-alpha.6
+
+* [FIXED] - make sure req.body is always an object
+* [FIXED] - throw an Error if a validator can not be found  with the provided key to `Route.prototype.validate`
+* [ADDED] - custom ajv validator keyword `$toJSON`
+* [ADDED] - `res.setPaginationHeaders` method
+* [FIXED] - set req.body to an empty object if req body parser did not set the property
+* [FIXED] - `requestContentType` parser should not require the parser function to be set, if no parser function is found, skip and continue request execution
+
 ## v1.0.0-alpha.5
 
 * [FIXED] - Important! `Route` uids were being incorrectly constructed when an App run under a basePath in addition to a hostname